Contest refactor
