+++ libs/shared/src/commonProtocol/index.ts
@@ -1,3 +1,4 @@
 export * from './chainConfig';
+export * from './contractHelpers/Contest';
 export * from './contractHelpers/Launchpad';
 export * from './utils';
